gpu: Move more debugprintf to be like gpuav
