+++ layers/gpu/spirv/module.h
@@ -68,7 +68,6 @@ class Module {
     // Order of functions that will try to be linked in
     std::vector<LinkInfo> link_info_;
     void LinkFunction(const LinkInfo& info);
-    bool IsInstrumented() const { return !link_info_.empty(); }
     void PostProcess();
 
     // The class is designed to be written out to a binary file.
